feat(用户服务): 扩展用户服务功能并修复事务装饰器
添加用户服务的获取和更新方法，完善用户端点逻辑
修复事务装饰器中未找到db session时的错误处理
更新健康检查接口以包含数据库连接验证
调整日志模型字段类型为Float
添加模块文档字符串和更新环境示例配置

+++ app/models/log.py
@@ -9,7 +9,7 @@
 import uuid
 from typing import Any
 
-from sqlalchemy import Boolean, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Float, ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -42,6 +42,6 @@ class OperationLog(AuditableModel):
     params: Mapped[Any | None] = mapped_column(JSONB, nullable=True)
     response_code: Mapped[int | None] = mapped_column(Integer, nullable=True)
     response_result: Mapped[Any | None] = mapped_column(JSONB, nullable=True)
-    duration: Mapped[float | None] = mapped_column(Integer, nullable=True)  # ms
+    duration: Mapped[float | None] = mapped_column(Float, nullable=True)
     ip: Mapped[str | None] = mapped_column(String(50), nullable=True)
     user_agent: Mapped[str | None] = mapped_column(String(500), nullable=True)
